pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipelinerl"
version = "0.1.0"
description = "A scalable asynchronous reinforcement learning implementation with in-flight weight updates."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "ServiceNow" },
]
dependencies = [
    "torch>=2.6",
    "vllm==0.8.3",
    "accelerate==1.7.0",
    "Tapeagents[finetune]==0.1.15",
    "transformers==4.51.0",
    "flash-attn==2.7.4.post1",
    "math-verify[antlr4_9_3]==0.7.0",
    "orjson==3.10.16",
    "redis==5.2.1",
    "hydra-core>=1.3.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pipelinerl*"]
